Revert request encoder protocol refactor from step 3
