+++ src/chipiron/environments/checkers/players/adapters/checkers_adapter.py
