+++ src/chipiron/players/factory_higher_level.py
@@ -23,7 +23,7 @@ from chipiron.players.observer_wiring import ObserverWiring
 from chipiron.players.player_args import PlayerFactoryArgs
 from chipiron.players.player_handle import InProcessPlayerHandle, PlayerHandle
 from chipiron.players.player_thread import PlayerProcess
-from chipiron.players.wirings.checkers_wiring import CHECKERS_WIRING
+from chipiron.environments.checkers.players.wiring.checkers_wiring import CHECKERS_WIRING
 from chipiron.players.wirings.chess_wiring import CHESS_WIRING
 from chipiron.utils.communication.mailbox import MainMailboxMessage
 from chipiron.utils.dataclass import IsDataclass
