+++ tests/checkers/test_checkers_recur_zipf_player_smoke.py
@@ -28,7 +28,7 @@ def test_checkers_tree_piececount_player_produces_parseable_legal_action_name() 
     from chipiron.players.move_selector.tree_and_value_args import TreeAndValueAppArgs
     from chipiron.players.player_args import PlayerFactoryArgs
     from chipiron.players.player_ids import PlayerConfigTag
-    from chipiron.players.wirings.checkers_wiring import (
+    from chipiron.environments.checkers.players.wiring.checkers_wiring import (
         BuildCheckersGamePlayerArgs,
         build_checkers_game_player,
     )
