+++ tests/checkers/test_checkers_terminal_scale.py
@@ -12,7 +12,7 @@ from chipiron.players.boardevaluators.evaluation_scale import (
     EvaluationScale,
     get_value_over_enum,
 )
-from chipiron.players.evaluators.checkers_piece_count import CheckersOverEventDetector
+from chipiron.environments.checkers.players.evaluators.checkers_piece_count import CheckersOverEventDetector
 
 
 @dataclass(frozen=True)
